feat: add CategoryCard and SearchLoading components; refactor HelpCenter and remove unused HeroSearch and SearchResults components
